Removed Sync + Send and their unsafe impl. For now we move those to the lamellar lamellae.
Refactored Info builder. Now hints is part of the builder pipeline itself instead of an argument.
Also, removed mutable access to Attribute objects from the public API. The only access allowed is to
query the attributes returned from getinfo in an read-only fashion.
These attributes structures have been refactored to use rust-native objects rather than the ones produced
by the C bindings, allowing us to remove pointers and other C-related types.

Introduced Version struct that can be used to define/retrieve the version of a fabric object or the library itself
in a high-level way, but can also interoperate with the C library.

Changed the implementation of `bind` method for Endpoints to enforce binding TX/RX/Shared(CQs/Counters)
in a single call per synchronization object.

 Refactored AV::insert flavors. All of them are now exposed through a single method that accepts
an enum with its variants denoting the flavor requested.
In AV introduced new constructors "with_name" and "read_only" that can be used instead of the respective methods

Refactored implementation of:
- IoVec
- Ioc
- Msg/MsgTagged/MsgRma
First, these structures now contain a lifetime which is used to track the lifetime of the buffers they point to.
Second, all of these types now have a `*Mut` flavor which explicitly captures the respective data by &mut
and denotes that these buffers will be mutated by the called their wrappers are passed to.

Removed the need to specify data type for the above structures (except IoC) so that they can provide the full benefits of the underlying library.

Changed the way Atomic/Collective operations map a rust datatype to the respective libfabric enum variant.
Instead of using TypeId::of::<T>()  the new implementation declares a new trait `AsFiType`which is implemented
for all types supported. This allows us to detect if the user is passing unsupported types at compile time
instead of runtime.